Add proper packaging
