+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"net"
 	"net/http"
@@ -9,102 +8,13 @@ import (
 	"os/signal"
 	"path/filepath"
 	"syscall"
-	"time"
 
+	"github.com/openebs/scope-plugin/metrics"
+	"github.com/openebs/scope-plugin/plugin"
 	log "github.com/sirupsen/logrus"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 )
-
-// URL is the address of cortex agent.
-const URL = "http://cortex-agent-service.maya-system.svc.cluster.local:80/api/v1/query?query="
-
-// Clientset contains kubernetes client..
-var (
-	ClientSet *kubernetes.Clientset
-)
-
-// PVMetrics groups the data a plugin needs.
-type PVMetrics struct {
-	ReadIops        float64
-	WriteIops       float64
-	ReadLatency     float64
-	WriteLatency    float64
-	ReadThroughput  float64
-	WriteThroughput float64
-}
-
-// Plugin struct groups the methods a plugin needs.
-type Plugin struct {
-	HostID string
-	Iops   map[string]PVMetrics
-}
-
-type sample struct {
-	Date  time.Time `json:"date"`
-	Value float64   `json:"value"`
-}
-
-type metric struct {
-	Samples []sample `json:"samples,omitempty"`
-	Min     float64  `json:"min"`
-	Max     float64  `json:"max"`
-}
-
-type metricTemplate struct {
-	ID       string  `json:"id"`
-	Label    string  `json:"label,omitempty"`
-	Format   string  `json:"format,omitempty"`
-	Priority float64 `json:"priority,omitempty"`
-}
-
-type node struct {
-	Metrics map[string]metric `json:"metrics"`
-}
-
-type topology struct {
-	Nodes           map[string]node           `json:"nodes"`
-	MetricTemplates map[string]metricTemplate `json:"metric_templates"`
-}
-
-type pluginSpec struct {
-	ID          string   `json:"id"`
-	Label       string   `json:"label"`
-	Description string   `json:"description,omitempty"`
-	Interfaces  []string `json:"interfaces"`
-	APIVersion  string   `json:"api_version,omitempty"`
-}
-
-type request struct {
-	NodeID string
-}
-
-type report struct {
-	PersistentVolume topology
-	Plugins          []pluginSpec
-}
-
-type response struct {
-	ShortcutReport *report `json:"shortcutReport,omitempty"`
-}
-
-// Metrics stores the json.
-type Metrics struct {
-	Status string `json:"status"`
-	Data   struct {
-		ResultType string `json:"resultType"`
-		Result     []struct {
-			Metric struct {
-				Name              string `json:"__name__"`
-				Instance          string `json:"instance"`
-				Job               string `json:"job"`
-				KubernetesPodName string `json:"kubernetes_pod_name"`
-				OpenebsPv         string `json:"openebs_pv"`
-			} `json:"metric"`
-			Value []interface{} `json:"value"`
-		} `json:"result"`
-	} `json:"data"`
-}
 
 func setupSocket(socketPath string) (net.Listener, error) {
 	os.RemoveAll(filepath.Dir(socketPath))
@@ -115,7 +25,6 @@ func setupSocket(socketPath string) (net.Listener, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to listen on %q: %v", socketPath, err)
 	}
-
 	log.Printf("Listening on: unix://%s", socketPath)
 	return listener, nil
 }
@@ -130,13 +39,6 @@ func setupSignals(socketPath string) {
 	}()
 }
 
-// Response unmarshal the obtained Metric json
-func Response(response []byte) (*Metrics, error) {
-	result := new(Metrics)
-	err := json.Unmarshal(response, &result)
-	return result, err
-}
-
 func main() {
 	// creating in-cluster config
 	config, err := rest.InClusterConfig()
@@ -145,7 +47,7 @@ func main() {
 	}
 
 	// create clientset of kubernetes
-	ClientSet, err = kubernetes.NewForConfig(config)
+	metrics.ClientSet, err = kubernetes.NewForConfig(config)
 	if err != nil {
 		log.Error(err)
 	}
@@ -165,182 +67,14 @@ func main() {
 		os.RemoveAll(filepath.Dir(socketPath))
 	}()
 
-	plugin := &Plugin{
+	plugin := &plugin.Plugin{
 		HostID: "host",
 	}
 
-	go plugin.updateIops()
+	go plugin.UpdateMetrics()
 
 	http.HandleFunc("/report", plugin.Report)
 	if err := http.Serve(listener, nil); err != nil {
 		log.Errorf("error: %v", err)
 	}
-}
-
-// Report is called by scope when a new report is needed. It is part of the
-// "reporter" interface, which all plugins must implement.
-func (p *Plugin) Report(w http.ResponseWriter, r *http.Request) {
-	rpt, err := p.makeReport()
-	if err != nil {
-		log.Errorf("error: %v", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	raw, err := json.Marshal(*rpt)
-	if err != nil {
-		log.Errorf("error: %v", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-	w.Write(raw)
-}
-
-func (p *Plugin) getPVTopology(PVName string) string {
-	return fmt.Sprintf("%s;<persistent_volume>", PVName)
-}
-
-func (p *Plugin) makeReport() (*report, error) {
-
-	metrics := make(map[string][]float64)
-	resource := make(map[string]node)
-
-	for pvUID, values := range p.Iops {
-		metrics[p.getPVTopology(pvUID)] = append(metrics[p.getPVTopology(pvUID)], values.ReadIops, values.WriteIops, values.ReadLatency, values.WriteLatency, values.ReadThroughput, values.WriteThroughput)
-	}
-
-	for pvName := range metrics {
-		resource[pvName] = node{
-			Metrics: p.metrics(metrics[pvName]),
-		}
-	}
-	rpt := &report{
-		PersistentVolume: topology{
-			Nodes:           resource,
-			MetricTemplates: p.metricTemplates(),
-		},
-		Plugins: []pluginSpec{
-			{
-				ID:          "openebs",
-				Label:       "OpenEBS Plugin",
-				Description: "Adds graphs of metrics of OpenEBS PV",
-				Interfaces:  []string{"reporter"},
-				APIVersion:  "1",
-			},
-		},
-	}
-	return rpt, nil
-}
-
-// Create the Metrics type on top-left side
-func (p *Plugin) metrics(data []float64) map[string]metric {
-	metrics := map[string]metric{
-		"readIops": {
-			Samples: []sample{
-				{
-					Date:  time.Now(),
-					Value: data[0],
-				},
-			},
-			Min: 0,
-			Max: 100,
-		},
-		"writeIops": {
-			Samples: []sample{
-				{
-					Date:  time.Now(),
-					Value: data[1],
-				},
-			},
-			Min: 0,
-			Max: 100,
-		},
-		"readLatency": {
-			Samples: []sample{
-				{
-					Date:  time.Now(),
-					Value: data[2],
-				},
-			},
-			Min: 0,
-			Max: 100,
-		},
-		"writeLatency": {
-			Samples: []sample{
-				{
-					Date:  time.Now(),
-					Value: data[3],
-				},
-			},
-			Min: 0,
-			Max: 100,
-		},
-		"readThroughput": {
-			Samples: []sample{
-				{
-					Date:  time.Now(),
-					Value: data[4],
-				},
-			},
-			Min: 0,
-			Max: 100,
-		},
-		"writeThroughput": {
-			Samples: []sample{
-				{
-					Date:  time.Now(),
-					Value: data[5],
-				},
-			},
-			Min: 0,
-			Max: 100,
-		},
-	}
-	return metrics
-}
-
-func (p *Plugin) metricTemplates() map[string]metricTemplate {
-	return map[string]metricTemplate{
-		"readIops": {
-			ID:       "readIops",
-			Label:    "Iops(R)",
-			Format:   "",
-			Priority: 0.1,
-		},
-		"writeIops": {
-			ID:       "writeIops",
-			Label:    "Iops(W)",
-			Format:   "",
-			Priority: 0.2,
-		},
-		"readLatency": {
-			ID:       "readLatency",
-			Label:    "Latency(R)",
-			Format:   "millisecond",
-			Priority: 0.3,
-		},
-		"writeLatency": {
-			ID:       "writeLatency",
-			Label:    "Latency(W)",
-			Format:   "millisecond",
-			Priority: 0.4,
-		},
-		"readThroughput": {
-			ID:       "readThroughput",
-			Label:    "Throughput(R)",
-			Format:   "bytes",
-			Priority: 0.5,
-		},
-		"writeThroughput": {
-			ID:       "writeThroughput",
-			Label:    "Throughput(W)",
-			Format:   "bytes",
-			Priority: 0.6,
-		},
-	}
-}
-
-func (p *Plugin) metricIDAndName() (string, string) {
-	return "OpenEBS Plugin", "OpenEBS Plugin"
 }
